feat/fix: função write_log_state
