+++ src/utils/genia/__init__.py
@@ -7,6 +7,9 @@ from .io_functions import (
     print_break_line,
     get_pretty_input,
     print_architecture,
+)
+
+from .log_functions import (
     get_log_path,
     write_log,
     write_log_state,
